# lib/openapi/generated.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RealEST Property Marketplace API",
    "description": "Nigerian property marketplace API with ML validation, geospatial search, and intelligent matching",
    "version": "2.0.0",
    "contact": {
      "name": "RealEST Support",
      "email": "support@realest.ng"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://realest.ng",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT Auth token"
      }
    },
...
  },
  "paths": {
...
    "/api/properties/{id}/favorites": {
      "post": {
        "method": "post",
        "summary": "Save property to favorites",
        "description": "Add a property to the authenticated user's saved/favorite properties. Property must be published.",
        "tags": [
          "Properties - Favorites"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Property ID"
          }
        ],
        "responses": {
          "201": {
            "description": "Property saved to favorites",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Property not found"
          },
          "409": {
            "description": "Property already saved to favorites"
          }
        },
        "operationId": "postPropertiesIdFavorites"
      },
      "delete": {
        "method": "delete",
        "summary": "Remove property from favorites",
        "description": "Remove a property from the authenticated user's saved/favorite properties.",
        "tags": [
          "Properties - Favorites"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Property ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Property removed from favorites",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Property not found in favorites"
          }
        },
        "operationId": "deletePropertiesIdFavorites"
      }
    },
...
    "/api/dashboard/owner/inquiries/{id}/respond": {
      "post": {
        "method": "post",
        "summary": "Respond to owner inquiry",
        "description": "Send a response to an inquiry and mark it as responded.",
        "tags": [
          "Dashboard - Owner Inquiries"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Inquiry ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "response_message"
                ],
                "properties": {
                  "response_message": {
                    "type": "string"
                  },
                  "contact_phone": {
                    "type": "string"
                  },
                  "contact_email": {
                    "type": "string",
                    "format": "email"
                  },
                  "schedule_viewing": {
                    "type": "boolean"
                  },
                  "viewing_date": {
                    "type": "string"
                  },
                  "viewing_time": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response sent successfully"
          },
          "400": {
            "description": "Invalid response data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied - Only property owners can respond to inquiries"
          },
          "404": {
            "description": "Inquiry not found"
          }
        },
        "operationId": "postDashboardOwnerInquiriesIdRespond"
      }
    },
...
    "/api/admin/validation/vetting/{id}": {
      "post": {
        "method": "post",
        "summary": "Update vetting action",
        "description": "Approve, reject, schedule, or flag a property in the vetting workflow.",
        "tags": [
          "Admin - Validation"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Property ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "action"
                ],
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "approve",
                      "reject",
                      "schedule",
                      "flag_issue"
                    ]
                  },
                  "notes": {
                    "type": "string"
                  },
                  "scheduled_date": {
                    "type": "string"
                  },
                  "rejection_reason": {
                    "type": "string"
                  },
                  "issue_details": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vetting action completed successfully"
          },
          "400": {
            "description": "Invalid request data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Admin access required"
          },
          "404": {
            "description": "Property not found or not eligible for vetting"
          }
        },
        "operationId": "postAdminValidationVettingId"
      }
    },
...
  },
...
}